Shard test/dist.js into parallel-safe spec files
Splits the monolithic 1769-line test/dist.js into:
- test/dist-runner.js: shared per-distribution test registration logic
  (registerDistributionTests), extracted so multiple spec files can reuse it
- test/dist-shard-0.js..dist-shard-3.js: four files that deterministically
  partition continuousCases.concat(discreteCases) by index modulo 4 and
  register their slice via the shared runner
- test/dist-base.js: the non-distribution-suite tests (base Distribution
  class, PreComputed, Degenerate/Kolmogorov special cases, parameter-count
  regressions, fit() precision gate) that were previously inlined in
  test/dist.js

Enables mocha --parallel (.mocharc.yml, package.json test script) to use
all available CPU cores. Verified: identical 5283 passing distribution
tests before and after the split (no cases dropped or duplicated), and
npm test (with nyc coverage) drops from >6m40s serial to ~2m27s parallel
on a 4-core machine, with coverage thresholds unchanged.

Resolves #1023
