+++ .mocharc.yml
@@ -1,2 +1,3 @@
 timeout: 120000
 recursive: true
+parallel: true
